🔧 chore(.gitignore): Remove unused entries and add missing newline at the end of the file
The `.gitignore` file was modified to remove unused entries and add a missing newline at the end of the file. The removed entries are:
- `docs`
The newline was added to comply with conventions.

🔧 chore(.husky): Remove unused husky configuration files

The `.husky/commit-msg`, `.husky/pre-commit`, and `.husky/pre-push` files were deleted as they were no longer needed. These files were used for configuring husky hooks for commit messages, pre-commit checks, and pre-push checks respectively. Since they were removed, it implies that these hooks are no longer used in the project.

🔧 chore(.prettierignore): Remove unused entry and add missing newline at the end of the file

The `.prettierignore` file was modified to remove an unused entry and add a missing newline at the end of the file. The removed entry is:
- `.yarn/`

💥 BREAKING CHANGE: Remove unused `.yarnrc.yml` file

The `.yarnrc.yml` file was deleted as it was no longer needed. This file contained configuration options for Yarn, including the compression level, node linker, plugins, yarn path, and changeset base refs. Since it was removed, it implies that these configurations are no longer used in the project. This is a breaking change as it may affect the behavior of Yarn in the project.

🔧 chore(package-lock.json): Remove unused `husky` dependency

The `husky` dependency was removed from the `package-lock.json` file as it was no longer used in the project.

🔧 chore(package.json): Remove unused `husky` dependency and postinstall script

The `husky` dependency was removed from the `package.json` file as it was no longer used in the project. Additionally, the `postinstall` script, which installed husky, was also removed.

These changes were made to optimize the project, remove unused configurations and dependencies, and ensure consistency in the codebase.